requirements.txt
gensim==3.8.3
pandas==1.1.5
numpy==1.18.5
progressbar33==2.4
utils==1.0.1
opencv_python==4.5.1.48
numpy==1.19.2
torchvision==0.9.1
six==1.15.0
torch==1.8.1
tqdm==4.60.0
chainer==7.7.0
matplotlib==3.4.2
nltk==3.6.2
pandas==1.2.4
Pillow==8.2.0
scikit_learn==0.24.2
scipy==1.6.3
sentence_transformers==1.2.0